Remove old, individual views for types of model selection
